Fixes #31 (Hangs on drawers) (#32)
org-wc.el (org-word-count-aux): Replace calls to org-end-of-meta-data
with explicit search for the end of the drawer, signalling an error if
it is not found.  Handle the case where we start in a drawer in the
same place.

The problem was that org-end-of-meta-data moves back to the heading
and then forward to the end of the metadata and any immediately
following drawers, causing an infinite loop before drawers that do not
immediately follow the metadata.